Restructure input processor modules.
