+++ src-tauri/src/toy_handling/input_processor/mod.rs
@@ -1,19 +1,9 @@
-pub mod penetration_systems;
-
+use dyn_clone::DynClone;
 use std::fmt::Debug;
 
-use dyn_clone::DynClone;
-use serde::{Deserialize, Serialize};
-use ts_rs::TS;
+use super::ModeProcessorInputType;
 
-use crate::frontend::{frontend_types::FePenetrationSystem, ToFrontend};
-
-use self::penetration_systems::PenetrationSystemType;
-
-use super::{
-    toyops::{ProcessingMode, ProcessingModeValues},
-    ModeProcessorInputType,
-};
+pub mod penetration_systems;
 
 /*
  * Penetration System Architecture
@@ -29,34 +19,3 @@ pub trait InputProcessor: DynClone + Debug + Send + Sync {
     fn is_parameter(&self, param: &String) -> bool;
 }
 dyn_clone::clone_trait_object!(InputProcessor);
-
-#[derive(Clone, Debug, Serialize, Deserialize, TS)]
-pub struct PenetrationSystem {
-    #[serde(skip)]
-    pub pen_system: Option<Box<dyn InputProcessor>>,
-    pub pen_system_type: PenetrationSystemType,
-    pub pen_system_processing_mode: ProcessingMode,
-    #[serde(skip)]
-    pub pen_system_processing_mode_values: ProcessingModeValues,
-}
-
-impl Default for PenetrationSystem {
-    fn default() -> Self {
-        Self {
-            pen_system: None,
-            pen_system_type: PenetrationSystemType::NONE,
-            pen_system_processing_mode: ProcessingMode::Raw,
-            pen_system_processing_mode_values: ProcessingModeValues::Raw,
-        }
-    }
-}
-
-impl ToFrontend<FePenetrationSystem> for PenetrationSystem {
-    type OutputType = FePenetrationSystem;
-    fn to_frontend(&self) -> Self::OutputType {
-        FePenetrationSystem {
-            pen_system_type: self.pen_system_type.clone(),
-            pen_system_processing_mode: self.pen_system_processing_mode.to_frontend(),
-        }
-    }
-}
